[Issue #6666] Fetch data for simpler GetApplicationZipRequest (#6774)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes for #ISSUE6666

## Changes proposed

<!-- What was added, updated, or removed in this PR. -->
Fetch data for simpler GetApplicationZipRequest

## Context for reviewers

<!-- Technical or background context, more in-depth details of the
implementation, and anything else you'd like reviewers to know about
that will help them understand the changes in the PR. -->
Now actually doing a query against the DB to get the file location on s3
to add the file information to the mtom response.

## Validation steps

<!-- Manual testing instructions, as well as any helpful references
(screenshots, GIF demos, code examples or output). -->