[Issue #6498] Setup shared schema and related utilities (#6727)
## Summary
Fixes #6498 

## Changes proposed
Add shared schemas for our form schemas (not yet referenced by the
forms)

Add a way to resolve a schema, removing all references (will be used in
subsequent PRs)

A bunch of utilities and tests

## Context for reviewers
The shared schemas that we want to define are going to work in the
following way:
* A schema will reference the schema by doing something like: `"$ref":
whatever_schema.field_ref("field")`
* When we write the data to our DB (and to a specific environment),
we'll resolve the schemas to not include refs
* Eventually we'll put these files on s3 as well in an unresolved
format.

The cool thing about this approach is that except for needing to do a
little bit of setup and configuration, it'll _just work_. Frontend will
need no changes, nothing else in our API outside of defining a schema
and where we send the schemas to an environment will need any changes,
and those'll be very minor.

If we didn't resolve the schema before writing it to places, a lot of
complexities result from needing any user of the API to be able to fetch
schemas from a specific location and then a lot of
timing/order-of-operations issues come up in the deployment. This just
makes that go away.

## Validation steps
All in tests so far, but I did a quick hack to verify that a fully
resolved schema in our DB works just fine with our frontend UI.