[Issue #6594] Setup applications for users/orgs in our local seed script (#6754)
## Summary
Fixes #6594

## Changes proposed
Setup a user with many applications in our local seed script

Refactor how we setup users/orgs to make it a bit less verbose/easier to
work with using the builder pattern

Adjust how competitions get setup to be reusable by the app creation
logic

## Context for reviewers
As we continue to add more test scenarios into our seed script, wanted
to make setting up users a bit simpler. Instead of having to define a
bunch of UUIDs to make everything line up properly, I hid away a lot of
that complexity into a builder utility that'll handle the setup. Should
be less verbose / more readable now.

For setting up applications, there are a few gaps in what I've done,
mainly that the way our auth is currently setup, for an org application
to work, you have to have permissions on the app AND org, which we'll be
changing soon. To make it so I could open the applications in our
frontend, I added both permissions, but left a note to remove some of
that later.

## Validation steps
A lot of running `make db-seed-local` and logging into the new user I
created and verifying the applications were setup properly.