[Issue #6593] Add opportunity information to application list schema (#6811)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes #6593 

## Changes proposed

<!-- What was added, updated, or removed in this PR. -->
Create `UserApplicationOpportunitySchema` and apply it to the
`UserApplicationCompetitionSchema`.
Send the opportunity data back in `get_user_applications`
Update tests 

## Context for reviewers

<!-- Technical or background context, more in-depth details of the
implementation, and anything else you'd like reviewers to know about
that will help them understand the changes in the PR. -->
In order to return the opportunity information (title, agency) needed in
the Application List view, the opportunity information needs to be
returned by the backend.

## Validation steps

<!-- Manual testing instructions, as well as any helpful references
(screenshots, GIF demos, code examples or output). -->
See updated unit tests

---------

Co-authored-by: nava-platform-bot <platform-admins@navapbc.com>