[Issue #6453] Make forms endpoints use new API key auth (#6820)
## Summary
Fixes #6453

## Changes proposed
* Modify the GET form and PUT form endpoints to use new API key auth
with RBAC
* Modify the list/update form scripts to be able to pass in the auth
* Add a new role for internal developers to use


## Context for reviewers
Because of the clunky way we need to do the two different approaches to
auth in update-forms, the auth logic is in the route as piping through
all the various types that the auth token can be would be a headache.
When we later remove the old auth approach, we can move it and simplify
a bunch of the logic / type issues.

## Validation steps
Will do some manual testing after merging in dev/staging once I've setup
my own personal user to have the roles.

---------

Co-authored-by: nava-platform-bot <platform-admins@navapbc.com>