[Issue #6662] Define schemas for GetSubmissionListExpanded (#6823)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes  #ISSUE6662 

## Changes proposed

<!-- What was added, updated, or removed in this PR. -->
Added schemas and response schemas for GetSubmissionListExpanded

## Context for reviewers

<!-- Technical or background context, more in-depth details of the
implementation, and anything else you'd like reviewers to know about
that will help them understand the changes in the PR. -->
This should allow us to consume the incoming xml on the request and
transform it into a python object, and take a response model and
transform it into a dictionary that can be turned into xml.

## Validation steps

<!-- Manual testing instructions, as well as any helpful references
(screenshots, GIF demos, code examples or output). -->