[Issue #6331] Add auto summation to sf424 and sf424a (#6451)
## Summary
Work for #6331

## Changes proposed
Add auto-summation to the SF424 and SF424A forms

## Context for reviewers
The actual implementation of auto-summation was already implemented, I
did fix one minor issue with rounding, but otherwise it just required a
whole bunch of configuration as there are 36 fields in the SF424A that
have to be summed.

## Validation steps
Need both the frontend and API running locally.

Run `make db-seed-local` to get form changes in your DB. Navigate to the
pilot-like opportunity with an application (via search or look at the
link the seed script gives you).

If you go to the 424 or 424a and populate any of the monetary fields,
the summations should work. The SF424A doesn't have the summed fields as
non-fillable yet (needs frontend changes due to how that form works),
but I was able to adjust the SF424:
<img width="674" height="825" alt="Screenshot 2025-09-19 at 1 30 28 PM"
src="https://github.com/user-attachments/assets/734eb926-7299-4565-9f4f-b9379fb8ae76"
/>


<img width="1197" height="685" alt="Screenshot 2025-09-19 at 1 32 51 PM"
src="https://github.com/user-attachments/assets/8f7350cc-fd80-4886-b3ec-87163541a092"
/>

---------

Co-authored-by: My <myduong@navapbc.com>