[Issue #6675] Create validation logic to verify an application that requires an org has an org (#6865)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes #6675 

## Changes proposed

<!-- What was added, updated, or removed in this PR. -->
Modify both our GET application endpoint and SUBMIT application logic to
check if organization is required but application doesn't have one. If
so, raise a new validation error.
Add new tests to cover the behavior.

## Context for reviewers

<!-- Technical or background context, more in-depth details of the
implementation, and anything else you'd like reviewers to know about
that will help them understand the changes in the PR. -->
We should verify that if the competition on an application requires
exactly an organization (not an individual), and no organization is on
the application, a validation issue is added.

## Validation steps

<!-- Manual testing instructions, as well as any helpful references
(screenshots, GIF demos, code examples or output). -->
See updated tests.