[ISSUE #6649] user invitation response endpoint (#6848)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes / Work for #6649  

## Changes proposed

Added endpoint `POST /v1/users/:user_id/invitations/{invitation_id}`
Added Request and Response Schema
Added Tests
## Context for reviewers

Allows a user to respond to an organization invitation by accepting or
rejecting it.
The endpoint validates:

- that the invitation exists, is still pending, and is not expired.
- ensures that the authenticated user matches the invitee’s ID and
email.
- ensures that the user's ID matches the invitee’s user ID.

On acceptance, the user is added to the organization with any linked
roles.