[Issue #6677] Create an endpoint for adding an organization to an application (#6881)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes #6677 

## Changes proposed

<!-- What was added, updated, or removed in this PR. -->
Create `PUT
/alpha/applications/:application_id/organizations/:organization_id`
endpoint
Allow that endpoint, with authz applied, to add an organization to an
application
Add supporting tests.

## Context for reviewers

<!-- Technical or background context, more in-depth details of the
implementation, and anything else you'd like reviewers to know about
that will help them understand the changes in the PR. -->
There are many new tests to cover the success/error scenarios based on
the below logic.

For auth:

User must have the Privilege.MODIFY_APPLICATION for the application

User must ALSO have Privilege.START_APPLICATION for the organization

There are many error scenarios like:
- The application already has an organization (422)
- The competition ONLY accepts individual applications (don’t want users
to be put in a bad spot) (422)
- The app/org don’t exist (404)
- User doesn’t have required privileges as described above (403)

Also some other business rules to consider when we add an organization
to an application we want to also do the following::

- Remove ALL application user records (in the future we might make this
a bit more refined, but this is necessary for now)
- Rerun the pre population logic so that values from the org get updated
into a users responses (like UEI) - similar to what we do in the
start-application logic (but on the already-existing forms, don't erase
a users answers).


## Validation steps

<!-- Manual testing instructions, as well as any helpful references
(screenshots, GIF demos, code examples or output). -->
See new / updated unit tests.