[Issue #6495] Switch json schemas to be fully resolved before usage (#6846)
## Summary
Fixes #6495

## Changes proposed
* Modifies our form schemas to be fully resolved in unit tests, our seed
script, and when reading/writing to the GET/PUT form endpoints
* Changes where we define the forms we want involved in seeding to be a
static list rather than reliant on fetching from the DB (keeps test data
from mixing with real data we want to work with)
* Remove DB session from form scripts as it's no longer needed with
above change

## Context for reviewers
This looks a lot bigger than it is, I changed all the form unit tests to
use the resolved form schema, which required a bit of piping through the
forms. I didn't change what we were actually testing in any of that
work.

We'll want to do some testing of the forms in lower envs once we update
them after this change just to verify that everything works as expected,
but some initial manual testing locally, I saw nothing changed with the
fully resolved form schemas (the frontend was already doing that).