[Issue #6908] Only fail the weekly vuln reporting job if something fails, not if it reports vulnerabilities (#6915)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes #6908  

## Changes proposed

Allow the option for the vuln scans to not fail job if they find
vulnerabilities

## Context for reviewers

When the job runs weekly, we bypass the ignore files so that we get a
full report of all fixed vulnerabilities, even if we're ignoring them.
But that would cause the job to fail as long as there were any active
ignores. This makes is so ignored vulnerabilities don't fail the job,
but an issue running the scanners would still fail the job.

## Validation steps

<!-- Manual testing instructions, as well as any helpful references
(screenshots, GIF demos, code examples or output). -->