[Issue #6496] Modify forms to use our shared schemas (#6859)
## Summary
Work for #6496  

## Changes proposed
* Changes all of our shared schemas to instead use the shared schemas
for common types (address, person name, monetary, etc.)

## Context for reviewers
This is the final step in getting us using our shared schemas, actually
referencing them. The prior PRs have already made it so the schemas
we're working with are fully resolved before using, and as we didn't
actually change the fully resolved schemas in this PR, everything just
works. All of our existing tests on the schemas to verify their behavior
did not require any changes.

Most fields previously defined in the `$defs` section of our schemas
were the shared schemas, so could be removed. A few forms still have a
few defs as they have types that only are needed in those particular
forms, so didn't need to be moved to the shared schema.

## Validation steps
Our tests for our schemas are thorough, if there were any issues with
this, the tests would be complaining. This _just_ works.