feat: adding the skeleton of the Applications page (#6907)
## Summary

Work for [#6595 ](https://github.com/HHS/simpler-grants-gov/issues/6595)

## Changes proposed

- [ ] Adding the skeleton for the Applications page at /applications
- [ ] Fixing the shifting h1 font on the Saved Opportunities, Saved
Searches, and API Dashboard page

## Context for reviewers

Will be followed up with
[#6597](https://github.com/HHS/simpler-grants-gov/issues/6597) to build
the table with data and the authentication check, currently the page is
accessible without logging in.

The shifting h1 fonts on the Saved Opportunities, Saved Searches, and
API Dashboard page was discovered when we standardizing the look across
similar Workspace pages. The font shifts smaller for tablet, then pops
bigger again for smaller breakpoints. Design asked to remove that,
apparently there had been an earlier initiative to do that and these
pages just got missed.

## Validation steps

- [ ] Go to http://localhost:3000/applications to see the skeleton page.
There is no link to it yet on the website.
- [ ] Check that the font looks ok at  all breakpoints on these pages: 
- [ ] Saved Opportunities as a logged in user at
http://localhost:3000/saved-opportunities
- [ ] Save Searches page as a logged in user at
http://localhost:3000/saved-search-queries
- [ ] API Dashboard page as a logged in user at
http://localhost:3000/api-dashboard (note that the odd spacing at the
top is current behaviour, this is just about the shifting of the font
size)