[Issue #6903] Add CD511 and form building docs (#6929)
## Summary
Fixes #6903 

## Changes proposed
* Adds the CD511 form
* Rewrote and expanded docs about how to build a form
* Moved a few common data fields (signature, date submitted, contact
person title) to the shared schema
* Added a quick UI schema - frontend folks will need to clean it up a
bit, but all the fields are there and functional. Not sure if they'll
want to adjust the big set of instructions at the top to be formatted
differently, I just kept them as rough paragraphs for each part.

## Context for reviewers
This form is pretty simple, you can see what it looks like for yourself
on https://www.grants.gov/forms/form-items-description/fid/276 - The
only complexity is that the `award_number` and `project_name` are both
required, but only as long as you haven't filled out the other one, or
in other words, at least one is required. The "right" way to implement
that in JSON schema would be to do a `oneOf` rule with two rules in it,
but the error that generates is vague and unhelpful (roughly "none of
the schemas in the oneOf validated" with no helpful details). Instead I
made two rules that both do a `if x is not included, y is required` and
vice-versa. This means that if neither is included, but are flagged, but
at least the error message is relatively intuitive.

I'm really curious what folks think of the docs I wrote, I need to
figure out where we'll put some of the form information that isn't
public that is easy for engineers to reach. Probably will look to put it
on our confluence?

## Validation steps
Easiest way to get to see the form:
* Run `make db-seed-local` - in the output you'll see a line early on
that says `Created a competition for form 'CD511 1.1'` with a link to an
opportunity with only this form. Go to that link
* Login as any user and start an app.

Here are some screenshots of what it looks like:
<img width="920" height="700" alt="Screenshot 2025-11-03 at 1 30 47 PM"
src="https://github.com/user-attachments/assets/73730826-e7e6-48a9-92de-121815ecb99a"
/>
<img width="515" height="729" alt="Screenshot 2025-11-03 at 1 30 52 PM"
src="https://github.com/user-attachments/assets/291673d8-5fa3-4dad-aa9d-211b91b2d5d4"
/>
<img width="489" height="716" alt="Screenshot 2025-11-03 at 1 30 58 PM"
src="https://github.com/user-attachments/assets/0ec7ee79-f3b8-441c-ba2f-0c20322b309d"
/>