[Issue #6779] Setup static competition/app in seed script (#6914)
## Summary
Fixes #6779

## Changes proposed
* Make a static opportunity/competition/application with all forms
* Removed logic that added the creator of an app to org-apps to mirror
our updated logic in the start-application logic

## Context for reviewers
The static app is meant for the frontend to have an easy application to
reach with all forms. If we add a new form to our form list, running
`make db-seed-local` WILL add that new form to the competition +
application.

## Validation steps
Tested that if not all the forms are present on the static
competition/app, they'll be added when running again (easiest way to get
that state is to change the `forms_to_add = forms` line to `forms_to_add
= forms[:3]` for a run to just add some of the forms).

If you login as the `many_app_user`, you should be able to go to the
application at this static URL:
http://localhost:3000/workspace/applications/application/892a6aa8-92b1-46c2-a26b-21a6aaab825d

<img width="1051" height="637" alt="Screenshot 2025-10-31 at 1 55 23 PM"
src="https://github.com/user-attachments/assets/ff4dfff7-dfe0-484a-bb9d-29f5d9763f9a"
/>
Note that a lot of the display values will differ as they're randomly
generated by our factory (award amounts, agency, etc.).