[Issue #6836] Make app attachments be soft-deleted (#6932)
## Summary
Fixes #6836

## Changes proposed
* Make application attachments be soft-deleted
* Make it so the `application_attachments` relationship list on an
application by default excludes the deleted attachments
* Test that the above relationship change works / is used everywhere

## Context for reviewers
In general, we want to move towards using soft-deletes as it simplifies
some data integrity issues in our system. In this case, we want to make
it so when we setup logic to add auditing to the application table, if
an attachment is deleted, we can continue to track that.

We still want to delete the actual file from s3, if someone deleted an
attachment because they accidentally uploaded the wrong file, we don't
want to hold onto it.

## Validation steps
Added tests to all places in the app endpoints that work with
application attachments to verify that deleted attachments aren't seen
by their logic.