[Issue #6834] Modify application submission logic to put XML file in the Zip (#6954)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes #6834

## Changes proposed

<!-- What was added, updated, or removed in this PR. -->
Add `enable_xml_generation` to `ApplicationSubmissionConfig`.
Add `SubmissionXMLAssembler` to handle assembling application submission
forms, header and footer into a `GrantApplication.xml` file, included in
the zip file.
When FF is enabled, `SubmissionXMLAssembler` will generate XML and
include in the ZIP file.
Add relevant tests to the task and assembler.
Update the submission manifest to reflect `GrantApplication.xml` is
added to the zip.

## Context for reviewers

<!-- Technical or background context, more in-depth details of the
implementation, and anything else you'd like reviewers to know about
that will help them understand the changes in the PR. -->
We moved up creating the `ApplicationSubmission` record earlier in the
`create_application_submission_task` because we need the
`legacy_tracking_number` in the XML footer, so we update the record
after creating it and the zip file.
The process is designed to *not* block the task if there was an error
generating the XML, so any forms that are not supported or fail are
skipped.


## Validation steps

<!-- Manual testing instructions, as well as any helpful references
(screenshots, GIF demos, code examples or output). -->
See unit tests for task and assembler.