[Issue #6808] Add tcertificates to TABLES_TO_LOAD (#6896)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes / Work for #6808 

## Changes proposed

<!-- What was added, updated, or removed in this PR. -->
Add "tcertificates" to TABLES_TO_LOAD to enable automated loading for
table.

## Context for reviewers

<!-- Technical or background context, more in-depth details of the
implementation, and anything else you'd like reviewers to know about
that will help them understand the changes in the PR. -->
This will allow tcertificates table to remain up to date with the Oracle
db. Waiting on #6807

## Validation steps

<!-- Manual testing instructions, as well as any helpful references
(screenshots, GIF demos, code examples or output). -->