[Unticketed] Change application list to sort by created_at by default (#7002)
## Summary

## Changes proposed
* Make the application list page sort by created_at instead of
updated_at
* Fix a SQLAlchemy warning in a recent change from the create submission
work

## Context for reviewers
After a recent chat with frontend/product, we agreed that sorting by
`created_at` makes more sense. `updated_at` is a bit clunky as it only
changes when a user updates the application name (form updates would not
change this). Created at is consistent as it gets set once.

The other fix is unrelated entirely, but fixes a SQLAlchemy warning that
was happening from an unrelated change.