moving the user/workspace page to dashboard (including text reference… (#7028)
…s) and adding applications links

## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes / Work for #6856 #6768 

## Changes proposed

* Rename the Workspace page to Activity Dashboard at `/dashboard`
* Add a section to point to the new Applications page
* List both pages appropriately in the header
* Remove caching for the `/applications` page

## Context for reviewers

We want the new Workspace page to be called the Activity Dashboard, and
the Applications page to be integrated with it. The dashboard is behind
the `userAdminOff` feature flag, so you'll need to make sure it is
disabled (?!?) in your environment.

I did no move/rename the workspace components directory because it
seemed to have a broader scope than just this page.

There is one reference to "Workspace" on the modal after saving a search
query. I left this reference as the link goes directly to the saved
searches page, which is under the "Workspace" header item.

## Validation steps

- [ ] The workspace page can now be found at `/dashboard`
- [ ] The title of the new tab is the Activity Dashboard 
- [ ] The breadcrumbs (which are likely to eventually be removed) list
the Activity Dashboard.
- [ ] The title of the new page is "Welcome to your activity dashboard."
Design requested we remove the color around Welcome.
- [ ] There is a new Applications card to the left of the "Saved
queries" card, and the link should take you to `/applications`.
- [ ] In the header the entry under their email to "Workspace" has been
removed.
- [ ] The link to "Activity Dashboard" now appears in the Workspace
header when the feature flag is disabled and the links are in
alphabetical order
- [ ] The link to "Applications" now appears in the Workspace header
regardless of feature flag status