[Issue #7011] Add Supplementary NEH Cover Sheet Form (#7056)
## Summary
Work for #7011

## Changes proposed
Add the Supplementary NEH Cover Sheet Form

## Context for reviewers
https://grants.gov/forms/form-items-description/fid/530

Another form, nothing too fancy about it although it has two fields that
one of them is required. Like the CD511, I just made it so they're both
required until one is set and then the other one will clear up.

## Validation steps
Can make sure the forms are loaded locally by doing `make db-migrate
db-seed-local`. Login as the as `many-app-user` and an app should
already be created which you can find at
http://localhost:3000/applications

Note that the form can't be displayed due to an issue that needs to be
fixed in the frontend for handling multiple if-statements in a nested
object. The frontend has created a ticket to address this issue.