Gblosser/5931 hide save button button only (#7005)
## Summary

Work for #5931

## Changes proposed

When the form is not in progress, the Save button is hidden
This PR includes work that will be implemented in a future PR for #5931

## Context for reviewers

This PR is part of larger work on the #5931 ticket, but _only_ features
the hidden save button change. The readonly changes are also present in
[this](https://github.com/HHS/simpler-grants-gov/pull/6870/) PR.

## Validation steps

After submitting an application, the Save button should not be visible
on forms.

---------

Co-authored-by: Glen Blosser <glenblosser@Glens-MacBook-Pro.local>