[Issue #7071] Filter out select application forms from XML generation (#7085)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes #7071 

## Changes proposed

<!-- What was added, updated, or removed in this PR. -->
Update XML assembler to apply the same rules as we do for application
submission for form inclusion.

## Context for reviewers

<!-- Technical or background context, more in-depth details of the
implementation, and anything else you'd like reviewers to know about
that will help them understand the changes in the PR. -->
Similar to application submission, confirm only required and forms
included in the submission are part of the XML generation process.

## Validation steps

<!-- Manual testing instructions, as well as any helpful references
(screenshots, GIF demos, code examples or output). -->
See new unit tests.