[Issue #6814] Create ecs task for setting up cert user (#7021)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes / Work for #6814 

## Changes proposed

<!-- What was added, updated, or removed in this PR. -->
Added a task/command to create a LegacyCertificate based on a
`tcertificate.currentcertid` and some `Role.role_ids`. Creates a `User`
of `UserType.LEGACY_CERTIFICATE` and assigns it to the agency associated
with the tcertificate along with adding the `AgencyUser` and
`AgencyUserRoles`.

```
make cmd args="task setup-cert-user --cert-id 1001 --role-ids <UUID> --role-ids <UUID>"
```

## Context for reviewers

<!-- Technical or background context, more in-depth details of the
implementation, and anything else you'd like reviewers to know about
that will help them understand the changes in the PR. -->
This will take a tcertificate and create an associated user with roles
attached to it.

## Validation steps

<!-- Manual testing instructions, as well as any helpful references
(screenshots, GIF demos, code examples or output). -->