[Issue #6861] Switch to supported version of the depreciated package (#6862)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Work for #6861  

## Changes proposed

Change from the old, depreciated non-namespaced package, to the new
namespaced one.

## Context for reviewers

This is reacting to a warning during the build process about the package
we had been referencing being depreciated

## Validation steps

Check out the branch, ensure that form functionality works with the more
current version of the package.