[Issue #6955] frontend adjustments to form cd511 (#7099)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Work for #6955 

## Changes proposed

Prettifying the layout for the Form CD511.

## Context for reviewers

Most of this work was previously done by @chouinar, this is just making
the large text section at the top more readable.

## Validation steps

A local opportunity linked to this form can be found by running `make
db-seed-local` and searching the output for "CD511".

- [ ] Successfully start an application against this form
- [ ] Ensure the form looks correct
- [ ] Ensure the form can be saved
- [ ] Ensure the application can be submitted
- [ ] Check the print setup to make sure that the PDF looks correct