[Issue #6949] Setup foreign staging tables legacy users[2/2] (#7052)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes for #6949  

## Changes proposed

Add `tuser_profile` and `vuser_account` tables to the load job for
hourly sync.

## Context for reviewers
Foreign and legacy tables have been successfully added in the staging
environment.
The load job was tested locally against the staging environment, and
data was confirmed to sync successfully.