[Issue #7234]: Hotfix for public files endpoint (#7235)
## Summary

The public files endpoint in all lower environments are not working. The
issue was we were applying to the same origin policy as the frontend and
the files endpoint. This was resulting in the aws authorization headers
not being passed correctly to the s3 origin.

## Validation steps

The public files download functionality should continue to work as is.