[Issue #6935]: adding the organizations page (#7230)
## Summary

Work for #6935 

## Changes proposed

We're creating a new Organizations page behind the `manageUsersOff`
feature flag so users can see the organizations where they are members,
and have easy access to link to the organization details and (if
applicable) manage the users in that organization.

## Context for reviewers

We'll be reusing a component already created on the Activity Dashboard
and updating it as well to have the Manage Users functionality.

I don't have unittests right now for any of the work gated by feature
flags due to the lack of examples in our code base for how to create
them. I've created a follow up ticket
[here](https://github.com/HHS/simpler-grants-gov/issues/7231).

## Validation steps

For setup, first set the feature flags
(http://localhost:3000/dev/feature-flags) page

- [ ] If manageUsersOff is set to enabled, then link should not appear
in header and when going to
[/organizations](http://localhost:3000/organizations) it should redirect
you to the maintenance page

For all other instructions, manageUsersOff is set to disabled

- [ ] If user is logged out when going to
[/organizations](http://localhost:3000/organizations) then user is
prompted to log in
- [ ] Organizations always appears in the header under workspace
alphabetically
- [ ] A user with no organizations (sign in with random new user name)
should see appropriate message on
[/organizations](http://localhost:3000/organizations) and
[/dashboard](http://localhost:3000/dashboard)
- [ ] A user with mixed organizations (sign in with `mixed_roles_user`)
sees two organizations
- [ ] The org where they are an admin they see the view details and the
manage users links
- [ ] The org where they are member they only see the view details links
- [ ] Links to view organization details are correct
- [ ] Links to the manage users page are correct 
- [ ] The Activity Dashboard Organizations section appears correct for
users as well
- [ ] In the case of error an appropriate error message is shown -- You
can do this by adding ` raise_flask_error(500, f"BREAKING API”)` to this
[line](https://github.com/HHS/simpler-grants-gov/blob/main/api/src/api/users/user_routes.py#L208)