[Unticketed] Schedule the build-automatic-opportunities job (#7208)
## Summary

## Changes proposed
Add a schedule to run a new command I've built for generating automatic
opportunities based on forms.

## Context for reviewers
This is a task that grabs every form in the environment, and generates
an opportunity if one doesn't yet exist for that form. Every time it
runs it always creates an opportunity that has all forms (that way as we
add forms, we don't need to adjust that, just find the latest).

I've manually run the task in each of dev/staging/training already. You
can see the opportunities this creates by searching `SGG` like so:
https://staging.simpler.grants.gov/search?query=SGG