[Issue #6815] Modify SOAP endpoint to enforce RBAC (#7149)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes / Work for #6815 

## Changes proposed

<!-- What was added, updated, or removed in this PR. -->
Add validation check for certificate. 

## Context for reviewers

<!-- Technical or background context, more in-depth details of the
implementation, and anything else you'd like reviewers to know about
that will help them understand the changes in the PR. -->
Using the serial_number on a certificate, get the LegacyCertificate,
which gives you the Agency and the User. Using the operation_name you
can get the correct Privilege. At that point you can check the User
access based on that Privilege. This should enforce RBAC.

- [x] Process the cert dnd pull the serial number out of it
- [x] Throw error If no cert is found
- [x] Throw error if cert is expired
- [x] Throw error if cert is not connected to agency (grantor only)
- [x] Figure out which privilege a user requires per operation
- [x] Calls RBAC code with the user associated with the cert, privilege
for operation, and agency

## Validation steps

<!-- Manual testing instructions, as well as any helpful references
(screenshots, GIF demos, code examples or output). -->