[Unticketed] Fix bug in jsonref resolver which incorrectly resolves schemas (#7334)
## Summary

## Changes proposed
* Ports a bugfix from the jsonref library in to fix an issue with
resolving schemas incorrectly.

## Context for reviewers
Added context in the ticket - but the issue this was causing was that
depending on the order we resolved schemas, it would sometimes
incorrectly resolve a schema to have the wrong fields. The linked fix
changes some internal method for how it tracks references to fix it, but
it hasn't been made into a release yet, so manually patched/overrode the
function to the newer fixed one.

## Validation steps
Before, running the tests in the forms folder could fail ( `make test
args="-x -s -vv tests/src/form_schema/forms"`).