[Issue  #7183] Change to align Instructions and submit button (#7280)
Summary
Fixes for https://github.com/HHS/simpler-grants-gov/issues/7183

Changes proposed
The "Submit Application" button and "Download Instructions" button are
not aligned on the Application page.

Context for reviewers
The "Submit Application" button and "Download Instructions" button are
not aligned on the Application page.
Aligning both buttons to appear on the same line for this i have updated
the bootstrap class to add more padding.

Validation steps
when submitting an application download application instructions and
submit application are not aligned. this has been updated