[Issue #6816] Modify soap proxy endpoints to use new env var (#7314)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes / Work for #6816 

## Changes proposed

<!-- What was added, updated, or removed in this PR. -->
We already have a environment variable called SOAP_AUTH_CONTENT that
we’ll want to make an alternative one. Instead of storing the ID /
private key, we just want a map from legacy certificate ID (a UUID) to
private key like so:

SOAP_PRIVATE_KEYS="{\"my-uuid\": \"-----BEGIN PRIVATE KEY ...\"}"

## Context for reviewers

<!-- Technical or background context, more in-depth details of the
implementation, and anything else you'd like reviewers to know about
that will help them understand the changes in the PR. -->
Now that we have a legacy_certificate table, we should be able to
associate a private key with a legacy_certificate_id instead of using a
fingerprint.

## Validation steps

<!-- Manual testing instructions, as well as any helpful references
(screenshots, GIF demos, code examples or output). -->